fix(auth): seed subscription codex defaults
