+++ src/utils/toml.ts
@@ -1,0 +1,84 @@
+export function readTopLevelTomlString(
+  content: string,
+  key: string,
+): string | null {
+  let inTopLevel = true;
+  const lines = content.split(/\r?\n/);
+  for (const line of lines) {
+    const trimmed = line.trim();
+    if (!trimmed || trimmed.startsWith("#")) continue;
+    if (/^\[[^[\]]+\]\s*(#.*)?$/.test(trimmed)) {
+      inTopLevel = false;
+      continue;
+    }
+    if (!inTopLevel) continue;
+    const match = line.match(/^\s*([A-Za-z0-9_.-]+)\s*=\s*(.*?)\s*(?:#.*)?$/);
+    if (!match || match[1] !== key) continue;
+    return parseTomlStringValue(match[2]);
+  }
+  return null;
+}
+
+export function upsertTopLevelTomlString(
+  content: string,
+  key: string,
+  value: string,
+): string {
+  const eol = content.includes("\r\n") ? "\r\n" : "\n";
+  const assignment = `${key} = "${escapeTomlString(value)}"`;
+
+  if (!content.trim()) {
+    return assignment + eol;
+  }
+
+  const lines = content.split(/\r?\n/);
+  let replaced = false;
+  let inTopLevel = true;
+
+  for (let i = 0; i < lines.length; i++) {
+    const line = lines[i];
+    const trimmed = line.trim();
+    if (!trimmed || trimmed.startsWith("#")) continue;
+    if (/^\[[^[\]]+\]\s*(#.*)?$/.test(trimmed)) {
+      inTopLevel = false;
+      continue;
+    }
+    if (!inTopLevel) continue;
+    const match = line.match(/^\s*([A-Za-z0-9_.-]+)\s*=/);
+    if (match && match[1] === key) {
+      lines[i] = assignment;
+      replaced = true;
+      break;
+    }
+  }
+
+  if (!replaced) {
+    const firstTableIndex = lines.findIndex((line) =>
+      /^\s*\[[^[\]]+\]\s*(#.*)?$/.test(line.trim()),
+    );
+    if (firstTableIndex >= 0) {
+      lines.splice(firstTableIndex, 0, assignment);
+    } else {
+      lines.push(assignment);
+    }
+  }
+
+  let out = lines.join(eol);
+  if (!out.endsWith(eol)) out += eol;
+  return out;
+}
+
+function parseTomlStringValue(value: string): string {
+  const trimmed = value.trim();
+  if (trimmed.startsWith('"') && trimmed.endsWith('"') && trimmed.length >= 2) {
+    return trimmed.slice(1, -1);
+  }
+  if (trimmed.startsWith("'") && trimmed.endsWith("'") && trimmed.length >= 2) {
+    return trimmed.slice(1, -1);
+  }
+  return trimmed;
+}
+
+export function escapeTomlString(value: string): string {
+  return value.replace(/\\/g, "\\\\").replace(/"/g, '\\"');
+}
